This patch is a trivial response to a good comment from Uwe, so I am self-acking before it gets lost.
Acked-by: Ronald G. Minnich <rminnich@gmail.com>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@2924 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/mainboard/pcengines/alix1c/cache_as_ram_auto.c
@@ -86,7 +86,7 @@ static u8 spdbytes[] = {
 	[SPD_tRRD] = 10,
 };
 
-static u8 spd_read_byte(unsigned device, unsigned address)
+static u8 spd_read_byte(u8 device, u8 address)
 {
 	print_debug("spd_read_byte dev ");
 	print_debug_hex8(device);
